Update testcase (more to come)
